Moving files into components folder
