+++ react/src/lib/components/ColorLegends/components/WebVizContinuousLegend.tsx
