+++ react/src/lib/components/ColorLegends/components/WebVizDiscreteLegend.tsx
