+++ react/src/lib/index.js
@@ -16,8 +16,8 @@ import VectorSelector from "./components/VectorSelector";
 import WellCompletions from "./components/WellCompletions";
 import { VectorCalculator } from "./components/VectorCalculator";
 import WellLogViewer from "./components/WellLogViewer";
-import WebVizContinuousLegend from "./components/ColorLegends/WebVizContinuousLegend";
-import WebVizDiscreteLegend from "./components/ColorLegends/WebVizDiscreteLegend";
+import WebVizContinuousLegend from "./components/ColorLegends/components/WebVizContinuousLegend";
+import WebVizDiscreteLegend from "./components/ColorLegends/components/WebVizDiscreteLegend";
 
 export {
     HistoryMatch,
